converter/requirements.txt
Django==5.1.1
djangorestframework==3.14.0
requests==2.30.0
